+++ inference.py
@@ -24,7 +24,7 @@ import torch
 import torch.nn.functional as F
 from torch.distributions import Categorical
 
-from src.app_config import ModelConfig
+from src.config import ModelConfig
 from src.envs.mario import MarioEnvConfig, MarioVectorEnvConfig, create_vector_env
 from src.models import MarioActorCritic